[fix] 파일 다운로드 동작하지 않는 문제와 다운로드 경로 수정

+++ app/src/main/java/com/tosslab/jandi/app/services/download/DownloadService.java
@@ -28,7 +28,7 @@ public class DownloadService extends IntentService implements DownloadController
     public static final String KEY_FILE_NAME = "file_name";
     public static final String KEY_FILE_EXTENSIONS = "ext";
     public static final String KEY_FILE_TYPE = "file_type";
-    public static final int NONE_FILE_ID = -1;
+    public static final long NONE_FILE_ID = -1;
     public static final String ACTION_STOP_DOWNLOAD_SERVICE = "com.tosslab.jandi.app.download.service.stop";
     public static final String EXTRA_STOP = "stop";
     public static final String EXTRA_NOTIFICATION_ID = "notification_id";
